refactor: 컴포넌트 잘게 쪼개기 완료
